maint: fixed convergerce issue for open systems
Now the fermi-level gets corrected in a better way using
estimations of the *true* Lorentzian integral.

This should also be ported to TS since it seems much better.
It however, only works when the charge is relatively close to the
correct value due to the use of the tan function.

Signed-off-by: Nick Papior <nickpapior@gmail.com>