bug: fixed wrong reference energy
Now there are never any shifts for the MFH. This allows
the reference energy to be initially specified by the user
through the Hamiltonian.

Another fix is the convergence rate of the Fermi level
in iterate3. Now we use the atan(x) trick for all energies to
converge Ef. It is faster but otherwise the same.

Lastly, with the reference energy fixed it enables strict checks for
the total energy. In this case I can get an exact agreement of the total
energy down to the 6th digit.

I have also removed the possibility to shift Ef in update_hamiltonian
as it shouldn't be used.

Signed-off-by: Nick Papior <nickpapior@gmail.com>