enh: enabled usage of pulay mixing from sisl
Now there is a basic implementation allowing Pulay
mixing. And indeed MANY iterations are saved.

This requires the latest sisl commit to function.

I have updated all codes in examples/* and checked
that they indeed improve efficiency.
I have however not checked all, so perhaps the
parameters needs changing for some of the examples.
Please consider double checking them.

Also, I have changed the convergence criteria.
Previously it was like this:

   dn = np.absolute(self.dm - ni).sum()

however, generally in DFT one does:

   dn = np.absolute(self.dm - ni).max()

The reason is that using sum, one will have a factor
of # of elements meaning that for large systems the criteria
will be stricter than for smaller systems. Instead one
typically looks at the largest change in *any* element.
This will also require you to go through the examples and
check tolerances etc. And probably also the defaults
in .converge. Note that 1e-10 is extremely low and
for Siesta the typical convergence criteria is of the order
1e-4 -- 1e-5. So you could probably do with 1e-6.
For an orthogonal basis this is equivalent to 1e-6 charge
difference.

Signed-off-by: Nick Papior <nickpapior@gmail.com>