enh: enabled SelfEnergy as direct input
Now NEGF class accepts list of SelfEnergy components
in combination with (HubbardHamiltonian, semi_inf).
This allows one to either use sisl.WideBandSE or
manually defined self-energies.

I think I have corrected all tests etc.

I think this is a bit cleaner and more general
in that the interface for NEGF need not be too complicated.

I have also reworked the bisection of the fermi-level in NEGF.
It should now converge a bit faster.

I have amended the changes done in #62 to replace it.