bug: fixed *stupid* bug
I had used [[None]*2]*3 which is not what it seems like.
Lists in Python are references to lists, and thus the *3
simply creates 3 lists with references to the same list.
The reason why this was working before (in master) was because
the arrays were created erroneously, but then casted to numpy.ndarray.
This meant that all values (None) were copied and thus the correct
decoupled elements were created.

Ensured such lists are created correctly now (using _nested_list).

Also updated the code for a better memory layout of the self-energies.
It now makes much more sense.